refactor: consolidate TOML domain validation into stringutil
manifest.Deserialize and the peer handshake Server-Domain check each
ported rippled's isProperlyFormedTomlDomain (StringUtilities.cpp:131-156)
independently, leaving two copies of one oracle that could drift. Extract
a single stringutil.IsProperlyFormedTomlDomain and call it from both.

Move the domain-predicate unit tests into the new package and add the
length-bound, 63/64-char label, and empty-label boundary cases.
