+++ internal/peermanagement/handshake_test.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"net"
 	"net/http"
-	"regexp"
 	"strconv"
 	"strings"
 	"testing"
@@ -1766,67 +1765,6 @@ func TestHeaderIPIsV6_FamilyFromTextForm(t *testing.T) {
 	assert.False(t, headerIPIsV6("1.2.3.4"))
 	assert.True(t, headerIPIsV6("2001:db8::1"))
 	assert.True(t, headerIPIsV6("::ffff:1.2.3.4"))
-}
-
-// isWellFormedDomain must agree with rippled's regex
-// (StringUtilities.cpp:142-153). Go regexp has no lookarounds, so the
-// no-leading/trailing-hyphen rule is encoded directly.
-func TestIsWellFormedDomain_RegexCrossCheck(t *testing.T) {
-	rippledLike := regexp.MustCompile(
-		`^([A-Za-z0-9](?:[A-Za-z0-9\-]{0,61}[A-Za-z0-9])?\.)+[A-Za-z]{2,63}$`,
-	)
-	check := func(s string) bool {
-		if len(s) < 4 || len(s) > 128 {
-			return false
-		}
-		return rippledLike.MatchString(s)
-	}
-
-	inputs := []string{
-		"a.io", "example.com", "validator.example.com", "node-1.example.org",
-		"a.b.c.d.example.com", "X-1.X-2.example.io",
-		"localhost", "example", "a.b", "x.123", "a.b.c",
-		"-bad.example.com", "bad-.example.com", "_bad.example.com",
-		"example.com.", "example..com", ".example.com",
-		"", "a",
-		strings.Repeat("a", 64) + ".example.com",
-		strings.Repeat("a", 200) + ".com",
-		"foo.bar.MUSEUM",
-	}
-	for _, s := range inputs {
-		want := check(s)
-		got := isWellFormedDomain(s)
-		assert.Equal(t, want, got, "isWellFormedDomain(%q): want=%v got=%v", s, want, got)
-	}
-}
-
-// isWellFormedDomain matches rippled's isProperlyFormedTomlDomain
-// (StringUtilities.cpp:131-156).
-func TestIsWellFormedDomain_TomlParity(t *testing.T) {
-	cases := []struct {
-		name string
-		in   string
-		want bool
-	}{
-		{"valid_two_label", "a.io", true},
-		{"valid_subdomain", "validator.example.com", true},
-		{"valid_with_digits", "node-1.example.org", true},
-		{"too_short", "a.b", false},
-		{"too_long", strings.Repeat("a", 120) + ".example.com", false},
-		{"single_label", "example", false},
-		{"numeric_tld", "x.123", false},
-		{"one_char_tld", "a.b.c", false},
-		{"trailing_dot", "example.com.", false},
-		{"leading_hyphen_label", "-bad.example.com", false},
-		{"trailing_hyphen_label", "bad-.example.com", false},
-		{"empty", "", false},
-		{"underscore", "bad_label.example.com", false},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.want, isWellFormedDomain(tc.in))
-		})
-	}
 }
 
 // WriteRawHandshakeRequest must place every issue-#270 header on the
